requirements.txt
# requirements.txt

matplotlib==3.8.4
numpy==1.26.4
pandas==2.2.2
PyYAML==6.0.1
# The following are dependencies of the packages above
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1